[REFACTOR] Modify Directory Structure.
  - Files would be placed under corresponding directory (Components, Modules, Util)
