+++ app/src/main/java/com/doyoonkim/knutice/di/components/MainActivityComponent.kt
@@ -3,14 +3,12 @@ package com.doyoonkim.knutice.di.components
 import android.app.Application
 import com.doyoonkim.bookmark.di.BookmarkModule
 import com.doyoonkim.common.di.CommonModule
-import com.doyoonkim.data.di.DataBindingModule
 import com.doyoonkim.data.di.DataModule
 import com.doyoonkim.domain.di.DomainModule
 import com.doyoonkim.knutice.MainActivity
-import com.doyoonkim.knutice.MainApplication
-import com.doyoonkim.knutice.di.AppModule
-import com.doyoonkim.knutice.di.DispatcherModule
-import com.doyoonkim.knutice.di.ViewModelFactoryModule
+import com.doyoonkim.knutice.di.modules.AppModule
+import com.doyoonkim.knutice.di.modules.DispatcherModule
+import com.doyoonkim.knutice.di.modules.ViewModelFactoryModule
 import com.doyoonkim.main.di.MainModule
 import com.doyoonkim.network.di.NetworkModule
 import com.doyoonkim.notification.di.NotificationBindingModule
@@ -23,7 +21,6 @@ import dagger.Component
     DispatcherModule::class,
     CommonModule::class,
     DataModule::class,
-    DataBindingModule::class,
     DomainModule::class,
     NetworkModule::class,
     NotificationModule::class,
